Decouple component imports from server (#4150)
